requirements.txt
streamlit
opencv-python
numpy
keras
scikit-learn
joblib
phonetics
pybase64
Pillow
tensorflow
